☁ Added a new Weather command
This is a command that retrieves and displays weather data for a specified city using the OpenWeatherMap API. It has a weather command that takes a city name as an argument and displays the weather data for the city in an embed message. It also includes an error handler for the chat event that displays an error message in an embed if a required argument is missing from the command. The bot token and API key and endpoint are stored in environment variables and accessed using the os.environ.get function.